+++ code/app/src/main/java/com/example/project_seekdeep/CalendarFragment.java
@@ -47,7 +47,7 @@ import java.util.List;
  * @author Nancy Lin
  *
  */
-public class CalendarFragment extends Fragment implements CalendarAdapter.OnItemListener, OnGetQueryDataListener{
+public class CalendarFragment extends Fragment implements CalendarAdapter.OnItemListener{
 
     private UserProfile loggedInUser;
     private TextView monthYearText;
@@ -55,6 +55,8 @@ public class CalendarFragment extends Fragment implements CalendarAdapter.OnItem
     private Date selectedDate;
     Calendar selectedDateCalendar;
     private LocalDate selectedLocalDate;
+
+    private OnGetQueryDataListener queryDataListener;
 
 
     /**
@@ -128,7 +130,6 @@ public class CalendarFragment extends Fragment implements CalendarAdapter.OnItem
 
         //get array to construct recycler
         CalendarAdapter.OnItemListener listener = this;
-        OnGetQueryDataListener queryDataListener = this;
 
 
         //construct recycler view
@@ -142,7 +143,7 @@ public class CalendarFragment extends Fragment implements CalendarAdapter.OnItem
             @Override
             public void onClick(View v) {
                 selectedDateCalendar.add(Calendar.MONTH, 1);
-                buildCalendar(listener, queryDataListener);
+                buildCalendar(listener);
             }
         });
 
@@ -151,7 +152,7 @@ public class CalendarFragment extends Fragment implements CalendarAdapter.OnItem
             @Override
             public void onClick(View v) {
                 selectedDateCalendar.add(Calendar.MONTH, -1);
-                buildCalendar(listener, queryDataListener);
+                buildCalendar(listener);
             }
         });
 
@@ -159,20 +160,38 @@ public class CalendarFragment extends Fragment implements CalendarAdapter.OnItem
 
     }
 
-    public void buildCalendar(CalendarAdapter.OnItemListener onItemListener, OnGetQueryDataListener queryDataListener){
+    public void buildCalendar(CalendarAdapter.OnItemListener onItemListener){
+
+
+
         monthYearText.setText(getMonthFromCalendar(  selectedDateCalendar ));
         Log.d("NANCY", "calendar time |" + selectedDateCalendar.getTime());
         ArrayList<String> daysInMonth = daysInMonthArray( selectedDateCalendar);
+        //ArrayList<String> moodsInMonth = createMoodsInMonthArray( daysInMonth, queryDataListener);
 
-        ArrayList<String> moodsInMonth = queryDataListener.onStartListener(daysInMonth);
+        new CalendarQueries().createMoodsInMonthArray(daysInMonth, selectedDateCalendar, loggedInUser, new OnGetQueryDataListener() {
+            @Override
+            public void onStart() {
+                //nothing
+            }
 
-        queryDataListener.onSuccess()
+            @Override
+            public void onSuccess(ArrayList<String> moodsInMonth) {
+                CalendarAdapter calendarAdapter = new CalendarAdapter(daysInMonth, moodsInMonth, onItemListener);
+                RecyclerView.LayoutManager layoutManager = new GridLayoutManager(requireContext(), 7);
+                calendarRecyclerView.setLayoutManager(layoutManager);
+                calendarRecyclerView.setAdapter(calendarAdapter);
 
-        ArrayList<String> moodsInMonth = createMoodsInMonthArray( daysInMonth, this);
-        CalendarAdapter calendarAdapter = new CalendarAdapter(daysInMonth, moodsInMonth, onItemListener);
-        RecyclerView.LayoutManager layoutManager = new GridLayoutManager(requireContext(), 7);
-        calendarRecyclerView.setLayoutManager(layoutManager);
-        calendarRecyclerView.setAdapter(calendarAdapter);
+                Log.d("NANCY", "On success");
+                calendarAdapter.notifyDataSetChanged();
+
+            }
+
+            @Override
+            public void onFailed(DatabaseError databaseError) {
+
+            }
+        });
     }
 
 
@@ -261,146 +280,6 @@ public class CalendarFragment extends Fragment implements CalendarAdapter.OnItem
         return  daysInMonthArray;
     }
 
-    private ArrayList<String> createMoodsInMonthArray(ArrayList<String> daysInMonthArray, OnGetQueryDataListener queryDataListener){
-        ArrayList<String> moodsInMonth = new ArrayList<>();
-        queryDataListener.onStartListener(daysInMonthArray);
-
-
-        CollectionReference moods = FirebaseFirestore.getInstance().collection("MoodDB");
-
-
-        int j = daysInMonthArray.size();
-        for (int i = 0; i < j; i++){
-            if (daysInMonthArray.get(i).isEmpty()){
-                moodsInMonth.add("");
-
-            } else{
-                /* Todo     query firebase for the mood
-                            Handle the empty days
-                          handle the ties in queries
-
-                 */
-
-                //Find  Date based on moodsInMonth
-
-                Calendar whatDay = (Calendar) selectedDateCalendar.clone();
-                whatDay.add(Calendar.DATE, -1*selectedDateCalendar.get(Calendar.DATE));
-                String dateNumberString = daysInMonthArray.get(i);
-                //date of the cell
-                whatDay.add(Calendar.DATE, Integer.parseInt(dateNumberString));
-                Date cellDate = whatDay.getTime();
-
-                ArrayList<DocumentSnapshot> queryReturn = new ArrayList<>();
-                Log.d("fuck",loggedInUser.getUsername());
-                moods.whereEqualTo("owner.username", loggedInUser.getUsername())
-                        //.whereLessThan("postedDate", new Date(cellDate.getTime()+86400000))
-                        //                        .whereGreaterThan("postedDate", cellDate);
-                        .get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                            @Override
-                            public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                                for(QueryDocumentSnapshot document : task.getResult()){
-                                    queryReturn.add(document);
-                                    Log.d("NANCY",  "snapshot doc: "+ document.toString());
-                                }
-
-                                List<Integer> largestEmotion = Arrays.asList(0,0,0,0,0,0,0,0);
-                                for (int k = 0; k< queryReturn.size()-1; k++){
-                                    String t = (String) queryReturn.get(k).get("emotionalState");
-                                    Log.d("Fuck me", t);
-                                    switch (t){
-                                        case "ANGER":
-                                            largestEmotion.set(0, largestEmotion.get(0)+1);
-                                            break;
-                                        case "CONFUSION":
-                                            largestEmotion.set(1, largestEmotion.get(1)+1);
-                                            break;
-                                        case "DISGUST":
-                                            largestEmotion.set(2, largestEmotion.get(2)+1);
-                                            break;
-                                        case "FEAR":
-                                            largestEmotion.set(3, largestEmotion.get(3)+1);
-                                            break;
-                                        case "HAPPINESS":
-                                            largestEmotion.set(4, largestEmotion.get(4)+1);
-                                            break;
-                                        case "SADNESS":
-                                            largestEmotion.set(5, largestEmotion.get(5)+1);
-                                            break;
-                                        case "SHAME":
-                                            largestEmotion.set(6, largestEmotion.get(6)+1);
-                                            break;
-                                        case "SURPRISE":
-                                            largestEmotion.set(7, largestEmotion.get(7)+1);
-                                            break;
-                                    }
-                                }
-                                int maxValue = Integer.MIN_VALUE;
-                                for (Integer integer: largestEmotion){
-                                    if (integer>maxValue)
-                                        maxValue=integer;
-                                }
-                                int loca = 0;
-                                loca = largestEmotion.indexOf(maxValue);
-                                String emotion = locaToEmotion(loca);
-                                moodsInMonth.add(emotion);
-                            }
-                        });
-            }
-
-
-        }
-
-        return moodsInMonth;
-    }
-    public String locaToEmotion(int loca){
-        String emotion = null;
-        switch (loca) {
-            //anger
-            case 1:
-                emotion = EmotionalStates.ANGER.getEmoticon();
-                break;
-
-            //confusion
-            case 2:
-                emotion = EmotionalStates.CONFUSION.getEmoticon();
-                break;
-
-            //disgust
-            case 3:
-                emotion = EmotionalStates.DISGUST.getEmoticon();
-                break;
-
-            //fear
-            case 4:
-                emotion = EmotionalStates.FEAR.getEmoticon();
-                break;
-
-            //happiness
-            case 5:
-                emotion = EmotionalStates.HAPPINESS.getEmoticon();
-                break;
-
-            //sadness
-            case 6:
-                emotion = EmotionalStates.SADNESS.getEmoticon();
-                break;
-
-            //shame
-            case 7:
-                emotion = EmotionalStates.SHAME.getEmoticon();
-                break;
-
-            //surprise
-            case 8:
-                emotion = EmotionalStates.SURPRISE.getEmoticon();
-                break;
-            default:
-                emotion = "";
-                break;
-        }
-
-        return  emotion;
-    }
 
 
 
@@ -415,105 +294,7 @@ public class CalendarFragment extends Fragment implements CalendarAdapter.OnItem
     }
 
 
-    @Override
-    public ArrayList<String> onStartListener(ArrayList<String> daysInMonthArray) {
-        ArrayList<String> moodsInMonth = new ArrayList<>();
-        CollectionReference moods = FirebaseFirestore.getInstance().collection("MoodDB");
 
 
-        int j = daysInMonthArray.size();
-        for (int i = 0; i < j; i++){
-            if (daysInMonthArray.get(i).isEmpty()){
-                moodsInMonth.add("");
 
-            } else{
-                /* Todo     query firebase for the mood
-                            Handle the empty days
-                          handle the ties in queries
-
-                 */
-
-                //Find  Date based on moodsInMonth
-
-                Calendar whatDay = (Calendar) selectedDateCalendar.clone();
-                whatDay.add(Calendar.DATE, -1*selectedDateCalendar.get(Calendar.DATE));
-                String dateNumberString = daysInMonthArray.get(i);
-                //date of the cell
-                whatDay.add(Calendar.DATE, Integer.parseInt(dateNumberString));
-                Date cellDate = whatDay.getTime();
-
-                ArrayList<DocumentSnapshot> queryReturn = new ArrayList<>();
-                Log.d("fuck",loggedInUser.getUsername());
-                moods.whereEqualTo("owner.username", loggedInUser.getUsername())
-                        //.whereLessThan("postedDate", new Date(cellDate.getTime()+86400000))
-                        //                        .whereGreaterThan("postedDate", cellDate);
-                        .get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                            @Override
-                            public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                                for(QueryDocumentSnapshot document : task.getResult()){
-                                    queryReturn.add(document);
-                                    Log.d("NANCY",  "snapshot doc: "+ document.toString());
-                                }
-
-                                List<Integer> largestEmotion = Arrays.asList(0,0,0,0,0,0,0,0);
-                                for (int k = 0; k< queryReturn.size()-1; k++){
-                                    String t = (String) queryReturn.get(k).get("emotionalState");
-                                    Log.d("Fuck me", t);
-                                    switch (t){
-                                        case "ANGER":
-                                            largestEmotion.set(0, largestEmotion.get(0)+1);
-                                            break;
-                                        case "CONFUSION":
-                                            largestEmotion.set(1, largestEmotion.get(1)+1);
-                                            break;
-                                        case "DISGUST":
-                                            largestEmotion.set(2, largestEmotion.get(2)+1);
-                                            break;
-                                        case "FEAR":
-                                            largestEmotion.set(3, largestEmotion.get(3)+1);
-                                            break;
-                                        case "HAPPINESS":
-                                            largestEmotion.set(4, largestEmotion.get(4)+1);
-                                            break;
-                                        case "SADNESS":
-                                            largestEmotion.set(5, largestEmotion.get(5)+1);
-                                            break;
-                                        case "SHAME":
-                                            largestEmotion.set(6, largestEmotion.get(6)+1);
-                                            break;
-                                        case "SURPRISE":
-                                            largestEmotion.set(7, largestEmotion.get(7)+1);
-                                            break;
-                                    }
-                                }
-                                int maxValue = Integer.MIN_VALUE;
-                                for (Integer integer: largestEmotion){
-                                    if (integer>maxValue)
-                                        maxValue=integer;
-                                }
-                                int loca = 0;
-                                loca = largestEmotion.indexOf(maxValue);
-                                String emotion = locaToEmotion(loca);
-                                moodsInMonth.add(emotion);
-                            }
-                        });
-            }
-
-
-        }
-
-
-        return moodsInMonth;
-
-    }
-
-    @Override
-    public String onSuccess(ArrayList<DocumentSnapshot> queryReturn) {
-
-    }
-
-    @Override
-    public void onFailed(DatabaseError databaseError) {
-
-    }
 }
