+++ code/app/src/main/java/com/example/project_seekdeep/OnGetQueryDataListener.java
@@ -1,14 +1,11 @@
 package com.example.project_seekdeep;
 
-import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
-import com.google.firebase.firestore.DocumentSnapshot;
 
 import java.util.ArrayList;
 
 public interface OnGetQueryDataListener {
-    public ArrayList<String> onStartListener(ArrayList<String> daysInMonthArray);
-
-    public String onSuccess(ArrayList<DocumentSnapshot> queryReturn);
+    public void onStart();
+    public void onSuccess(ArrayList<String> moodsInMonth);
     public void onFailed(DatabaseError databaseError);
 }
